requirements.txt
numpy>=1.19.0,<2.0
torch>=1.13.0
torchvision>=0.14.1
Pillow>=8.0.0
scipy>=1.7.0
matplotlib>=3.3.0
thop>=0.1.0

